Refactor: re-organize around, do newtype wrappers
